Load 3.2.2 into safir_sdk_core/trunk.
git-svn-id: http://localhost:9111/private/safir_open/safir_sdk_core/trunk@4 9dd7cf02-b8c9-4ff8-a4a5-e7fda3bdd973

+++ src/dose/dose_communication.ss/src/dose_com/IpmSocket.h
@@ -23,7 +23,7 @@
 ******************************************************************************/
 
 #ifdef _LINUX
-#define INVALID_SOCKET   -1
+#define INVALID_SOCKET (unsigned long)-1
 #define SOCKET_ERROR    -1
 typedef  unsigned long SOCKET;
 typedef unsigned long IPADDR;
